+++ examples/v2/createPoolComposableStable.ts
@@ -62,7 +62,7 @@ const createPoolComposableStable = async (): Promise<{
         exemptFromYieldProtocolFeeFlag: false,
         swapFee: '0.01',
         poolOwnerAddress: signerAddress, // Balancer DAO Multisig,
-        balancerVersion: 2,
+        vaultVersion: 2,
         chainId,
     };
     const { call, to } = createPool.buildCall(createPoolComposableStableInput);