+++ test/v2/initPool/weighted.integration.test.ts
@@ -67,7 +67,7 @@ describe('Add Liquidity Init - Weighted Pool', async () => {
             swapFee: '0.01',
             poolOwnerAddress: signerAddress, // Balancer DAO Multisig
             chainId,
-            balancerVersion: 2,
+            vaultVersion: 2,
         };
 
         createTxInput = {