+++ test/v2/removeLiquidity/composableStable.recovery.integration.test.ts
@@ -126,7 +126,7 @@ class MockApi {
             address: testPool.address,
             type: testPool.type,
             tokens,
-            balancerVersion: 2,
+            vaultVersion: 2,
         };
     }
 }